refactored Reservation Card code
- separated reservation card logic into custom hook
- moved CalendarDatePicker and CalendarTimePicker into their own components with their own styles
- implemented a couple more edge cases: checking if there reserved dates exist, if the date select exists inside the reservedTimes hashmap
- made default times an empty string
- set values of reservation fields for form submission
